feat: use Metro Symlink resolver and cleanup our custom resolver
