--- tools/metro-extra-config/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@ledgerhq/metro-extra-config",
-  "version": "1.0.1",
-  "description": "Extra helpers for configuring the metro bundler",
-  "main": "index.js",
-  "private": true,
-  "dependencies": {
-    "@rnx-kit/metro-config": "^1.3.2",
-    "@rnx-kit/metro-resolver-symlinks": "0.1.32",
-    "metro-config": "^0.80.0",
-    "metro-resolver": "^0.80.0"
-  },
-  "author": "",
-  "license": "ISC"
-}
